[TLS] Allow setting SNI and ALPN.
There a are two new options now that allow setting the SNI hostname and
the ALPN protocol to 'nats' in the ClientHello TLS handshake message.
The SNI host name allows to connect a nats server behind a TLS
terminating proxy like traefik. Although the ALPN extension is not
strictly necessary for that, it makes it even more flexible to configure
proxy rules.